requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
python-dateutil==2.8.2
requests==2.31.0
aiohttp==3.9.1
gunicorn==21.2.0
aiohttp-worker==0.0.2